alerts/our_projects/lambdas/notifyfailures/requirements.txt
requests==2.18.4
pyyaml==3.12
